README.md
This readme will be updated within a couple of weeks. This is bah shell script for the Linux Terminal; I developed this to the default server installations directorys and to the local host machine. This script will continue gorw and become a more secured easily used script for the most commonly used task and a complete man.